+++ src/App.tsx
@@ -1,58 +1,30 @@
-/* -------------------------------------------------------------------------- */
-/*                                   Imports                                  */
-/* -------------------------------------------------------------------------- */
-
+import React, { useEffect, useState, useMemo } from "react";
+import { getAuth } from "firebase/auth";
+import { Routes, Route, Navigate, Outlet } from "react-router-dom";
 import { Box, useToast } from "@chakra-ui/react";
-import React, { useEffect, useState, useCallback, useMemo } from "react";
-import { connect } from "react-redux";
-import { mapStateToProps, mapDispatchToProps, login } from "./utility/Redux";
-import { Routes, Route, Navigate, useNavigate, Outlet } from "react-router-dom";
-
-import "firebase/database";
-import "firebase/compat/database";
-import { ref, child, get, onValue } from "firebase/database";
-
-import { auth, db, fetchData } from "./utility/Firebase";
-import Login from "./pages/Login";
-import { showErrorToast } from "./utility/ShowToast";
-
-import ChatRoom from "./pages/ChatRoom";
-import Register from "./pages/Register";
-import {
-  ContextWrapper,
-  USER_CONTEXT_DEFAULT,
-} from "./components/ContextWrapper";
-
-/* -------------------------------------------------------------------------- */
-/*                             Main App Component                             */
-/* -------------------------------------------------------------------------- */
+import { Login, ChatRoom, Register } from "./pages";
+import { ContextWrapper } from "./components";
+import { fetchData, showErrorToast } from "./utility";
+import { UserType } from "./types";
 
 export function App() {
-  const stateUser = useState(USER_CONTEXT_DEFAULT);
+  const stateUser = useState<UserType>();
   const [user, setUser] = stateUser;
+  const auth = getAuth();
 
   const toast = useToast();
   const toastIdRef = React.useRef();
 
-  /* --------------------- Authentication Functionalities --------------------- */
-
-  /* Reference:
-   * https://stackoverflow.com/a/69869761
-   */
-  const PrivateRoute = useMemo(() => {
-    let ls_user = localStorage.getItem("disclone-user");
-    let ls_uid = localStorage.getItem("disclone-uid");
-
-    if (user === null && ls_uid !== null) {
-      login(JSON.parse(ls_user as any), ls_uid);
-    }
-
-    return ls_user ? <Outlet /> : <Navigate to="/login" />;
-  }, [user]);
+  const PrivateRoute = useMemo(
+    () => (auth.currentUser || user ? <Outlet /> : <Navigate to="/login" />),
+    [auth.currentUser, user]
+  );
 
   useEffect(() => {
     auth.onAuthStateChanged((authUser) => {
       if (!authUser) return;
+
+      console.log(authUser);
 
       fetchData(`user/${authUser.uid}`)
         .then((result) => {
@@ -65,7 +37,7 @@ export function App() {
           showErrorToast(toast, toastIdRef);
         });
     });
-  }, [setUser, toast]);
+  }, [auth, setUser, toast]);
 
   return (
     <ContextWrapper stateUser={stateUser}>
