--- src/components/Layout/ChatRoom/index.js
@@ -1,2 +1,0 @@
-export * from "./ChatContainer";
-export * from "./ChatRoom";
