+++ src/components/Layout/Side/index.ts
@@ -1,0 +1,7 @@
+export * from "./Side";
+export * from "./SideChannel";
+export * from "./SideChannelDescription";
+export * from "./SideChannelList";
+export * from "./SideChannelMember";
+export * from "./SideHeading";
+export * from "./SideNewChannel";
