+++ src/components/Layout/index.ts
@@ -1,0 +1,2 @@
+export * from "./ChatRoom";
+export * from "./Side";
