--- src/components/Main.tsx
@@ -1,300 +1,0 @@
-/* -------------------------------------------------------------------------- */
-/*                                   Imports                                  */
-/* -------------------------------------------------------------------------- */
-
-import { Box, Text, useToast, IconButton } from "@chakra-ui/react";
-
-import React, {
-  useState,
-  useEffect,
-  useCallback,
-  useMemo,
-  useRef,
-} from "react";
-import { connect } from "react-redux";
-import { mapStateToProps, mapDispatchToProps } from "../utility/Redux";
-import { MdMenu } from "react-icons/md";
-
-import { writeData, fetchData, db } from "../utility/Firebase";
-import firebase from "firebase/compat/app";
-
-import InfiniteScroll from "react-infinite-scroll-component";
-import { Fragment } from "./Fragment";
-import SendChat from "./SendChat";
-import { showErrorToast } from "../utility/ShowToast";
-import { increment, ref, update } from "firebase/database";
-import { ChannelType, MessageType, StateType, UserType } from "../types";
-
-/* -------------------------------------------------------------------------- */
-/*              Contains the fragments (all chats) in a channel.              */
-/* -------------------------------------------------------------------------- */
-
-const range = 10;
-
-const ChatFragments = ({ channel, chats = [] }: any) => {
-  /* Reference (for the infinite scroll):
-   * https://blog.logrocket.com/4-ways-to-render-large-lists-in-react/
-   */
-
-  const [count, setCount] = useState({
-    prev: Math.max(chats.length - range, 0),
-    next: chats.length,
-  });
-  const [hasMore, setHasMore] = useState(true);
-  const [current, setCurrent] = useState(chats.slice(count.prev, count.next));
-  const initialized = useRef(undefined);
-
-  const getMoreData = useCallback(() => {
-    const floor = Math.max(count.prev - range, 0);
-
-    setTimeout(() => {
-      let newData = chats.slice(floor, count.prev - 1);
-
-      setCurrent((prev) =>
-        newData.concat(
-          prev.filter((old) => !newData.some((d) => d.id === old.id))
-        )
-      );
-    }, 1000);
-    setCount({
-      prev: floor,
-      next: count.next,
-    });
-    if (floor === 0) {
-      setHasMore(false);
-    }
-  }, [count.prev, count.next, chats]);
-
-  /* Reference
-   * https://stackoverflow.com/questions/19614069/get-percentage-scrolled-of-an-element-with-jquery
-   *
-   * Used to detect when the element is at top, such that older chats are rendered, if there is any.
-   */
-  const checkAndGetMoreData = useCallback(() => {
-    const root = document.getElementById("scrollable");
-
-    if (!root) return;
-
-    let scrollPercentage =
-      (100 * root.scrollTop) / (-root.scrollHeight + root.clientHeight);
-
-    console.log("Scrolling");
-
-    if (hasMore && scrollPercentage >= 99) {
-      getMoreData();
-    }
-  }, [hasMore, getMoreData]);
-
-  useEffect(() => {
-    let element = document.getElementById("scrollable");
-    element.onscroll = () => checkAndGetMoreData();
-  }, [checkAndGetMoreData]);
-
-  useEffect(() => {
-    setCurrent(chats.slice(Math.max(chats.length - range, 0), chats.length));
-    setHasMore(true);
-  }, [channel, chats]);
-
-  console.log(current);
-
-  return (
-    <InfiniteScroll
-      hasChildren={true}
-      endMessage="Already End"
-      loader={<></>}
-      dataLength={current.length}
-      inverse={true}
-      next={getMoreData}
-      hasMore={hasMore}
-      scrollThreshold="10px"
-    >
-      <Box
-        position="relative"
-        paddingTop="3.2rem"
-        paddingBottom="7.5rem"
-        zIndex="-1"
-      >
-        {current.map((value, idx) => (
-          <Fragment key={`chat-${channel}-${value.id}`} data={value} />
-        ))}
-      </Box>
-    </InfiniteScroll>
-  );
-};
-
-/* -------------------------------------------------------------------------- */
-/*                               Main Component                               */
-/* -------------------------------------------------------------------------- */
-
-interface MainProps {
-  channel?: ChannelType;
-  user: UserType;
-  stateChats: StateType<MessageType[]>;
-  stateSide: StateType<boolean>;
-}
-
-export function Main({
-  user,
-  channel,
-  stateChats,
-  stateSide,
-}: // chats = [],
-// user,
-// channel,
-// channels,
-// userId,
-// drawer,
-// openDrawer,
-// closeDrawer,
-MainProps) {
-  const toast = useToast();
-  const toastIdRef = React.useRef();
-
-  const [side, setSide] = stateSide;
-  const [chats, setChats] = stateChats;
-  const cannotSendChat = channel?.property.isReadOnly;
-  const { id: userId } = user;
-
-  const writeChat = useCallback(
-    (id, message) => {
-      writeData(`message/${channel.id}/${id}`, message);
-    },
-    [channel]
-  );
-
-  const handleSendChat = useCallback(
-    async (text) => {
-      let newId = await fetchData("counter/");
-
-      if (cannotSendChat) return;
-
-      let message = {
-        id: newId.message,
-        author: userId,
-        message: text,
-        timestamp: new Date().toLocaleString(),
-      };
-
-      console.log("ID!!");
-
-      const updates: any = {};
-      updates[`counter/message`] = increment(1);
-      try {
-        await update(ref(db), updates);
-        writeData(`message/${channel.id}/${message.id}`, message);
-      } catch (e) {
-        showErrorToast(toast, toastIdRef);
-      }
-
-      // database
-      //   .ref("counter/")
-      //   .child("message")
-      //   .set(firebase.database.ServerValue.increment(1))
-      //   .then(() => {
-      //     writeChat(id.message, message);
-      //   })
-      //   .catch((e) => {
-      //     showErrorToast(toast, toastIdRef);
-      //   });
-    },
-    [cannotSendChat, channel.id, toast, userId]
-  );
-
-  const renderMask = useMemo(() => {
-    return (
-      <Box
-        id="mask"
-        display={side ? "unset" : "none"}
-        position="fixed"
-        width="100vw"
-        height="100vh"
-        top="0"
-        left="0"
-        zIndex="2"
-        background="rgba(0.2, 0.2, 0.2, 0.4)"
-        onClick={() => setSide((prev) => !prev)}
-      ></Box>
-    );
-  }, [setSide, side]);
-
-  const renderChat = useMemo(
-    () => <ChatFragments channel={channel} chats={chats} />,
-    [channel, chats]
-  );
-
-  const renderInputChat = useMemo(
-    () => (
-      <Box
-        display="flex"
-        position="fixed"
-        height="3.2rem"
-        top="0"
-        right="0"
-        paddingLeft="4rem"
-        alignItems="center"
-      >
-        <Box
-          display="flex"
-          width={mainWidth}
-          position="relative"
-          justifyContent="flex-start"
-          paddingLeft="4rem"
-          alignItems="center"
-          height="3.2rem"
-          zIndex="4"
-          bg="#252329"
-          shadow="md"
-        >
-          <IconButton
-            aria-label="Toggle Drawer"
-            display={{ base: "flex", lg: "none" }}
-            colorScheme="gray"
-            icon={<MdMenu />}
-            minWidth="0"
-            width="2rem"
-            height="2rem"
-            marginRight="2rem"
-            onClick={() => setSide((prev) => !prev)}
-          />
-          <Text
-            lineHeight="1rem"
-            fontSize="1.2rem"
-            fontWeight="bold"
-            fontFamily="Noto Sans"
-          >
-            {channel?.name ?? "Unknown Channel"}
-          </Text>
-        </Box>
-      </Box>
-    ),
-    [channel?.name, setSide]
-  );
-
-  return (
-    <Box
-      display="flex"
-      flexDirection="column-reverse"
-      position="fixed"
-      width={mainWidth}
-      height="100vh"
-      top="0"
-      right="0"
-      paddingLeft="4rem"
-      paddingRight="4rem"
-      bg="#252329"
-      overflowY="auto"
-      id="scrollable"
-    >
-      {renderMask}
-      {renderChat}
-      {renderInputChat}
-      <SendChat
-        onSendChat={handleSendChat}
-        isDisabled={cannotSendChat}
-        // mainWidth={mainWidth}
-      />
-    </Box>
-  );
-}
-
-const mainWidth = { base: "100vw", lg: "calc(100vw - 384px)" };
