--- src/components/index.js
@@ -1,2 +1,0 @@
-export * from "./Auth";
-export * from "./Form";
