+++ src/pages/ChatRoom.tsx
@@ -1,19 +1,12 @@
-import React, {
-  useState,
-  useEffect,
-  useCallback,
-  useRef,
-  useMemo,
-} from "react";
-import { Side } from "../components/Side";
-import { Main } from "../components/Main";
+import { useState, useEffect, useRef, useMemo } from "react";
 import { onValue, ref } from "firebase/database";
+import { useToast } from "@chakra-ui/react";
+import { Side, ChatRoomContainer } from "../components";
+import { useUser } from "../hooks";
 import { db } from "../utility";
 import { ChannelType, MessageType } from "../types";
-import { useToast } from "@chakra-ui/react";
-import { useUser } from "../hooks";
 
-const ChatRoom = () => {
+export function ChatRoom() {
   const stateSide = useState(false);
   const stateChannels = useState<Record<number, ChannelType>>({});
   const [channels, setChannels] = stateChannels;
@@ -42,28 +35,6 @@ const ChatRoom = () => {
 
   const [user, setUser] = useUser();
 
-  const toast = useToast();
-  const toastIdRef = useRef();
-  // const history = useNavigate();
-
-  // Force dark mode, I can't figure out the Chakra UI method to do this.
-  // if (localStorage.getItem("chakra-ui-color-mode") !== "dark") {
-  //   localStorage.setItem("chakra-ui-color-mode", "dark");
-  //   history("/");
-  // }
-
-  // const getChannels = useCallback(() => {
-  //   onValue(ref(db, `channel/`), (snapshot) => {
-  //     if (snapshot.exists()) {
-  //       const data = snapshot.val();
-  //       console.log("Get Channels:");
-  //       console.log(data);
-  //       downloadChannel(data);
-  //     }
-  //   });
-  //   setInit(true);
-  // }, [downloadChannel]);
-
   useEffect(() => {
     onValue(ref(db, `channel/`), (snapshot) => {
       if (snapshot.exists()) {
@@ -74,44 +45,6 @@ const ChatRoom = () => {
       }
     });
   }, [init, setChannels]);
-
-  // const initialize = useCallback(async () => {
-  //   if (init || user === null) return;
-
-  //   let rawData = await fetchData(`channel/`);
-  //   console.log(rawData);
-  //   downloadChannel(rawData);
-  //   setChannel(0);
-
-  //   getChannels();
-  // }, [downloadChannel, getChannels, init, setChannel, user]);
-
-  // const handleSwitchChannelData = useCallback(async () => {
-  //   let members = channels[channel].member;
-  //   let memberOfChannel = {};
-
-  //   for await (const [memberId, status] of Object.entries(members)) {
-  //     let rawData = await fetchData(`user/${memberId}/`);
-  //     if (rawData !== null) {
-  //       memberOfChannel = {
-  //         ...memberOfChannel,
-  //         [memberId]: rawData,
-  //       };
-  //       memberOfChannel[memberId] = rawData;
-  //     }
-  //   }
-
-  //   const stringified = JSON.stringify(memberOfChannel);
-  //   usersChannel(stringified);
-  // }, [channel, channels, usersChannel]);
-
-  // useEffect(() => {
-  //   if (!channel || !channels || !channels[channel]) {
-  //     return;
-  //   }
-
-  //   handleSwitchChannelData();
-  // }, [channel, channels, chatChannel, handleSwitchChannelData]);
 
   useEffect(() => {
     onValue(ref(db, `message/${selectedChannel}/`), (snapshot) => {
@@ -124,10 +57,6 @@ const ChatRoom = () => {
     });
   }, [selectedChannel, setMessages]);
 
-  // useEffect(() => {
-  //   initialize();
-  // }, [initialize]);
-
   const renderContent = useMemo(
     () => (
       <>
@@ -136,7 +65,7 @@ const ChatRoom = () => {
           stateChannels={stateChannels}
           stateSelectedChannel={stateSelectedChannel}
         />
-        <Main
+        <ChatRoomContainer
           stateChats={stateMessages}
           stateSide={stateSide}
           user={user}
@@ -155,6 +84,4 @@ const ChatRoom = () => {
   );
 
   return <div>{init && renderContent}</div>;
-};
-
-export default ChatRoom;
+}
