+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export * from "./ChatRoom";
+export * from "./Login";
+export * from "./Register";
