--- src/utility/index.js
@@ -1,5 +1,0 @@
-export * from "./Firebase";
-export * from "./Initials";
-export * from "./Redux";
-export * from "./ShowToast";
-export * from "./Validation";
